feat: share bilingual segmentation across providers
